refact action handlers
