--- src/webcli2/action_handlers/config/__init__.py
@@ -1,1 +1,0 @@
-from .main import ConfigHandler
