--- src/webcli2/action_handlers/mermaid/__init__.py
@@ -1,1 +1,0 @@
-from .main import MermaidHandler
